feat: add GPU screen recorder, harden security, absorb kimi-code and modularize LKM / eBPF map registry
Integrated SovereignScreenRecorder inspired by Bandicam with GPU acceleration, target FPS pacing, pre-record buffering, dynamic bitrate scaling, multi-track audio, overlays, and live HUD. Hardened security by exposing SELinux, AppArmor, defensive audits, intrusion monitor, and zeroization. Absorbed next-gen code agent architecture (KimiCodeAgent) with Model Context Protocol (MCP) tool registrations, parallel subagents (Coder, Explorer, Planner), safety hooks, and Agent Client Protocol (ACP) JSON-RPC standard. Fully expanded StandardsComplianceManager to enforce POSIX, FHS, LSB, System V IPC, and UEFI GOP splash validation protocols. Modularized Linux-inspired eBPF Map state-sharing structures and Dynamic Kernel Module (LKM) Loader with GPL license taint verifications. Fixed integration test imports and lcov coverage gate.

+++ src/filesystem/support.rs
@@ -59,7 +59,7 @@ impl Filesystem for SimpleFilesystem {
         self.id
     }
     fn fs_type(&self) -> FilesystemType {
-        unsafe { core::mem::transmute(self.fs_type.load(Ordering::SeqCst) as u32) }
+        unsafe { core::mem::transmute(self.fs_type.load(Ordering::SeqCst)) }
     }
 
     fn mount(&mut self, _device: &[u8], mountpoint: &[u8]) -> Result<(), FilesystemError> {
